+++ include/slang/ast/types/AllTypes.h
@@ -163,11 +163,11 @@ class SLANG_EXPORT FixedSizeUnpackedArrayType : public Type {
 public:
     const Type& elementType;
     ConstantRange range;
-    uint32_t selectableWidth;
+    uint64_t selectableWidth;
     uint32_t bitstreamWidth;
 
     FixedSizeUnpackedArrayType(const Type& elementType, ConstantRange range,
-                               uint32_t selectableWidth, uint32_t bitstreamWidth);
+                               uint64_t selectableWidth, uint32_t bitstreamWidth);
 
     static const Type& fromDims(const Scope& scope, const Type& elementType,
                                 std::span<const ConstantRange> dimensions,
@@ -253,7 +253,7 @@ public:
 class SLANG_EXPORT UnpackedStructType : public Type, public Scope {
 public:
     std::span<const FieldSymbol* const> fields;
-    uint32_t selectableWidth = 0;
+    uint64_t selectableWidth = 0;
     uint32_t bitstreamWidth = 0;
     int systemId;
 
@@ -288,7 +288,7 @@ public:
 class SLANG_EXPORT UnpackedUnionType : public Type, public Scope {
 public:
     std::span<const FieldSymbol* const> fields;
-    uint32_t selectableWidth = 0;
+    uint64_t selectableWidth = 0;
     uint32_t bitstreamWidth = 0;
     int systemId;
     bool isTagged;
